refactor: move lib/social to ui/links for user-facing link management
