+++ ui/profile/ProfileCard.tsx
@@ -12,11 +12,11 @@ import ProfileAvatar from "@/ui/profile/ProfileAvatar";
 import useProfileLinks from "@/ui/profile/useProfileLinks";
 import VerifyProfileModal from "@/ui/verification/VerifyProfileModal";
 import { RedirectModal } from "@/ui/profile/editorModals";
-import { connectSocial } from "@/lib/social/connect";
-import { useConnectCallback } from "@/lib/social/useConnectCallback";
-import { upsertVerifiedLink } from "@/lib/social/verifyLink";
-import { detectProviderFromUrl } from "@/lib/social/avatars";
-import { PROVIDERS } from "@/lib/social/providers";
+import { connectSocial } from "@/ui/links/connect";
+import { useConnectCallback } from "@/ui/links/useConnectCallback";
+import { upsertVerifiedLink } from "@/ui/links/verifyLink";
+import { detectProviderFromUrl } from "@/ui/links/avatars";
+import { PROVIDERS } from "@/ui/links/providers";
 import { enrichLink } from "@/lib/profile/profileLinks";
 import ProfileCardListView from "./ProfileCardListView";
 import ProfileCardActions from "./ProfileCardActions";
